Fix: Separate functions in 2 archives for better structure understanding
