+++ src/functions/deck.js
@@ -11,21 +11,6 @@ export function createDeck() {
 	return deck;
 };
 
-export function shuffle(array) {
-  let currentIndex = array.length;
-  const shuffledArray = [...array];
-
-  while (currentIndex !== 0) {
-    const randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex--;
-
-    [shuffledArray[currentIndex], shuffledArray[randomIndex]] = [
-      shuffledArray[randomIndex], shuffledArray[currentIndex]];
-  }
-  
-  return shuffledArray;
-};
-
 export function sliceDeckInThree(deckParam) {
   const deck = deckParam;
   
@@ -66,7 +51,7 @@ export function switchStacks(deckToSwitch, stackName, initialStack, middleStack,
     default:
       break;
   }
-  
+
   return deckToSwitch;
 }
